+++ Data/ExportedWorldEvents.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-11-11 11-29-51 ]] --
+-- [[ Exported at 2022-11-16 19-54-29 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --
 
@@ -168,22 +168,22 @@ function exportedWorldEvents.LoadCategories(e)
     e[7299].Category = addon.Tabs["Expansions"].Categories[14].Children[7]; -- Primal Storms
     e[7300].Category = addon.Tabs["Expansions"].Categories[14].Children[7]; -- Primal Storms
     e[7301].Category = addon.Tabs["Expansions"].Categories[14].Children[7]; -- Primal Storms
-    e[7366].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[22]; -- Badlands
-    e[7367].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[22]; -- Badlands
-    e[7368].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[22]; -- Badlands
-    e[7369].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[22]; -- Badlands
-    e[7370].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[9]; -- Northern Barrens
-    e[7371].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[9]; -- Northern Barrens
-    e[7372].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[9]; -- Northern Barrens
-    e[7373].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[9]; -- Northern Barrens
-    e[7374].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[5]; -- Durotar
-    e[7375].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[5]; -- Durotar
-    e[7376].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[5]; -- Durotar
-    e[7377].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[5]; -- Durotar
-    e[7378].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[21]; -- Un'Goro Crater
-    e[7379].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[21]; -- Un'Goro Crater
-    e[7380].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[21]; -- Un'Goro Crater
-    e[7381].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[21]; -- Un'Goro Crater
+    e[7366].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[20]; -- Badlands
+    e[7367].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[20]; -- Badlands
+    e[7368].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[20]; -- Badlands
+    e[7369].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[20]; -- Badlands
+    e[7370].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[8]; -- Northern Barrens
+    e[7371].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[8]; -- Northern Barrens
+    e[7372].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[8]; -- Northern Barrens
+    e[7373].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[8]; -- Northern Barrens
+    e[7374].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[6]; -- Tirisfal Glades
+    e[7375].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[6]; -- Tirisfal Glades
+    e[7376].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[6]; -- Tirisfal Glades
+    e[7377].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[1].Children[6]; -- Tirisfal Glades
+    e[7378].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[20]; -- Un'Goro Crater
+    e[7379].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[20]; -- Un'Goro Crater
+    e[7380].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[20]; -- Un'Goro Crater
+    e[7381].Category = addon.Tabs["Expansions"].Categories[5].Children[1].Children[2].Children[20]; -- Un'Goro Crater
 end
 
 function exportedWorldEvents.InjectOptions()