+++ Localization/enUS.ExtraWrathStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2022-11-15 17-20-22 ]] --
+-- [[ Exported at 2022-11-18 08-21-57 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --